separated testing of metrics with dynamic tags
